fix(architecture): break source control import cycle
Pre-commit hook ran. Total eslint: 0, total circular: 0
