+++ src/store/workstation/codeEditor/sourceControlScopeAtom.ts
@@ -1,6 +1,6 @@
 import { atom } from "jotai";
 
-import type { SourceControlScopeMap } from "@src/modules/WorkStation/CodeEditor/Panels/EditorPrimarySidebar/tabs/sourceControlScopePickerHelpers";
+import type { SourceControlScopeMap } from "./sourceControlTypes";
 
 /** Per-repo Source Control scope for the current app session (not persisted). */
 export const sourceControlScopeMapAtom = atom<SourceControlScopeMap>({});
